requirements.txt
fastapi==0.88.0
numpy==1.23.3
pandas==1.5.0
pydantic==1.10.2
scikit_learn==1.1.3
uvicorn==0.20.0
neptune-client
